+++ packages/orq-rc/src/models/components/index.ts
@@ -3,5 +3,4 @@
  */
 
 export * from "./deployments.js";
-export * from "./publiccontact.js";
 export * from "./security.js";