+++ packages/orq-rc/RELEASES.md
@@ -3539,3 +3539,13 @@ Based on:
 - [typescript v3.12.0-rc.49] packages/orq-rc
 ### Releases
 - [NPM v3.12.0-rc.49] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.49 - packages/orq-rc
+
+## 2025-09-05 20:07:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.611.0 (2.694.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.12.0-rc.50] packages/orq-rc
+### Releases
+- [NPM v3.12.0-rc.50] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.50 - packages/orq-rc